tests/generated/test_menu.py
#!/usr/bin/env python3
"""Auto-generated test for: Menu
Category: Mixpanel | Quadrant: Q3 - Test Second | Risk: 36 (I:3 x P:4 x D:3)
TODO: Replace placeholder selectors."""
import pytest
import allure

@allure.suite("Mixpanel")
@allure.sub_suite("Menu")
@allure.tag("Q3")
@pytest.mark.q3
class TestMenu:
    """Q3 - Test Second tests for Menu."""

    @allure.title("Menu - Screen Launch")
    @allure.severity(allure.severity_level.NORMAL)
    def test_screen_launches(self, driver):
        pass

    @allure.title("Menu - Basic Functionality")
    @allure.severity(allure.severity_level.NORMAL)
    def test_basic_functionality(self, driver):
        pass

    @allure.title("Menu - Visual Regression")
    @allure.severity(allure.severity_level.NORMAL)
    @pytest.mark.visual
    def test_visual_regression(self, driver, visual_check):
        # TODO: Navigate then call visual_check("menu")
        pass
