requirements.txt
# PDR-I QA Toolkit v2.1 - Mac Mini M1 Compatible
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
streamlit>=1.32
plotly>=5.18
scikit-learn>=1.4
Pillow>=10.0
Appium-Python-Client>=4.0
pytest>=8.0
allure-pytest>=2.13
